+++ src/main/java/br/gov/servicos/arrecadacao/DebitoAutomaticoMovimentoRepositorio.java
@@ -14,7 +14,7 @@ public class DebitoAutomaticoMovimentoRepositorio {
 	@PersistenceContext
 	private EntityManager entity;
 
-	public void apagarMovimentosDebitoAutomaticoDasConta(List<Long> ids){
+	public void apagarMovimentosDebitoAutomaticoDasConta(List<Integer> ids){
 		String delete = "delete from arrecadacao.debito_auto_movimento where cnta_id in (:ids)";
 		
 		entity.createNativeQuery(delete)
