+++ src/main/java/br/gov/servicos/faturamento/ContaGeralRepositorio.java
@@ -13,13 +13,13 @@ public class ContaGeralRepositorio {
 	@PersistenceContext
 	private EntityManager entity;
 
-	public Long salvar(ContaGeral contaGeral) {
+	public Integer salvar(ContaGeral contaGeral) {
 		entity.persist(contaGeral);
 		entity.flush();
 		return contaGeral.getId();
 	}
 	
-	public void alterarHistoricoParaContasDeletadasPorReprocessamento(List<Long> idsContas){
+	public void alterarHistoricoParaContasDeletadasPorReprocessamento(List<Integer> idsContas){
 		String atualizar = "update faturamento.conta_geral set cntg_ichistorico = 3 where cnta_id in (:ids)";
 		
 		entity.createNativeQuery(atualizar)
