+++ src/test/java/br/gov/servicos/faturamento/ContaImpostosDeduzidosRepositorioTest.java
@@ -34,8 +34,8 @@ public class ContaImpostosDeduzidosRepositorioTest {
 	@ShouldMatchDataSet("conta_impostos_deduzidos_expected.yml")
 	@Cleanup(phase = TestExecutionPhase.AFTER, strategy = CleanupStrategy.USED_ROWS_ONLY)
 	public void apagarImpostosDeduzidos(){
-		List<Long> ids = new ArrayList<Long>();
-		ids.add(1L);
+		List<Integer> ids = new ArrayList<Integer>();
+		ids.add(1);
 		repositorio.apagarImpostosDeduzidosDeContas(ids);
 	}
 }
